Adds optional SRDF controller_box collision shapes
This change adds an optional controller_box collision
shape to be in sync with the URDF from sawyer_description.

This can be disabled, by invoking with controller_box:=false
when invoking $roslaunch sawyer_moveit_config sawyer_moveit.launch

It is set to true by default in order to prevent accidental
trajectory executions through the controller box.